+++ examples/reason-react-example/src/Opaque.re.js
@@ -2,7 +2,7 @@
 
 const OpaqueBS = require("./Opaque.bs");
 
-// No need to import locally visible type t. Make sure it is also marked with @genFlow
+// No need to import locally visible type t. Make sure it is also marked with @genType
 
 export opaque type t = any;
 export const fromInt: (number) => t = OpaqueBS.fromInt;