+++ examples/reason-react-example/src/basics/Component2.re.js
@@ -3,7 +3,7 @@
 const Component2BS = require("./Component2.bs");
 const CreateBucklescriptBlock = require("bs-platform/lib/js/block.js");
 
-// No need to import locally visible type block. Make sure it is also marked with @genFlow
+// No need to import locally visible type block. Make sure it is also marked with @genType
 
 export opaque type VariantA = any;
 export const A: VariantA = 0;