+++ examples/typescript-example/src/Records.tsx
@@ -3,9 +3,9 @@
 // tslint:disable-next-line:no-var-requires
 const RecordsBS = require("./Records.bs");
 
-// No need to import locally visible type business. Make sure it is also marked with @genFlow
-// No need to import locally visible type coord. Make sure it is also marked with @genFlow
-// No need to import locally visible type person. Make sure it is also marked with @genFlow
+// No need to import locally visible type business. Make sure it is also marked with @genType
+// No need to import locally visible type coord. Make sure it is also marked with @genType
+// No need to import locally visible type person. Make sure it is also marked with @genType
 
 import {list} from '../src/shims/ReasonPervasives.shim';