+++ Changes.md
@@ -1,5 +1,7 @@
 # master
-  Add "export default ComponentName" for TypeScript.
+  [Add "export default ComponentName" for TypeScript](https://github.com/cristianoc/genFlow/issues/21).
+  
+  [Fix: props are contravariant](https://github.com/cristianoc/genFlow/issues/22).
 
 # 0.7.0
   Print log information to .genFlowLog in the project root insted of standard output.